partially revert "Support of write command and request in bgapi"
commit 6c0e21c60889b528d48b6797960433a4e0211312 upstream

I don't know what the code is doing exaclty here, however since the
said commit characteristic write with bgapi backend is broken here.

The said commit implements an alternate code path when
wait_for_response=True, that was previously not implemented and it
also changes the code in case wait_for_response=False.

This commit does not touch the new code for the True case; on the
other hand, in case wait_for_respose=False, we stick to the old code
that here still works.

Signed-off-by: Andrea Merello <andrea.merello@gmail.com>