Don't reset Bluegiga dongles on open
Resetting the dongle ensures that everything is clean, so it looks like a
good idea at a first glance, however it turned out to be fairly nasty for
a bunch of reasons:

- This takes time. And every OS is different, so timeouts/retries must
  be properly tuned.

- Resetting the dongle causes the USB device to go down and then up
  again. If we use the dongle from inside a VM this has to be taken into
  account otherwise the USB device is lost.

- I have seen cases in which after a noticeable number of resets the
  dongle fails to come up again - no usb - on both win/linux. This is a
  real PITA.

To address this issues we don't reset the dongle anymore.

Other mechanisms have been added to try to clean up things as much as
possible:

- We do open the serial port with zero timeout and we perform a huge read
  in order to drain the RX fifo, then we close the serial port and we and
  open it again with proper timeout.

- We iterate on all possible handles to force a disconnect on each one in
  order to make sure that any possible zombie connection is killed.

NOTE: resetting the dongle would require to be able to send valid commands
to the dongle; this means that in case the TX queue in messed up for any reason
then the reset itself wouldn't work, so we don't mind about such cases (if ever
happens): removing reset doesn't make any difference (we would be screwed up
in either cases).